+++ sorbet/tapioca/require.rb
@@ -1,12 +1,15 @@
 # typed: true
 # frozen_string_literal: true
 
-require 'base64'
 require 'fileutils'
 require 'json'
 require 'multi_json'
 require 'openssl'
-require 'rbnacl'
+begin
+  require 'rbnacl'
+rescue LoadError
+  nil
+end
 require 'securerandom'
 require 'simplecov'
 require 'sorbet-runtime'